+++ src/main/java/mate/academy/demo/service/CategoryServiceImpl.java
@@ -37,9 +37,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
-        categoryRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Can't find category with id: " + id)
-        );
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find category with id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
